Remove useless network requests, change names
