+++ app/src/main/java/com/h/pixeldroid/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.h.pixeldroid
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.View
+import android.widget.Button
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val button = findViewById<Button>(R.id.button)
+        button.setOnClickListener((View.OnClickListener {
+            val intent = Intent(this, ProfileActivity::class.java)
+            startActivity(intent) }))
+    }
+}
